Utils: Fix big number string parsing error
Current configuration parsing logic can't handle big number strings so a
string is just left as a string without getting converted to an integer
and eventually it introduces crash issues in dbus-sensors because the
string can't be cast to an integer directly. This issue could be observed
when we add an I3C device which has 0x4cc31020000 as its Address setting
as an example.

To avoid this issue, this commit replaces 'stoi' and 'stoul' to 'stoll'
and 'stoull' respectively.

Tested: Big numbers were parsed properly.

+++ src/Utils.cpp
@@ -340,7 +340,7 @@ std::optional<std::string> templateCharReplace(
                     int constant = 0;
                     try
                     {
-                        constant = std::stoi(*it);
+                        constant = std::stoll(*it);
                     }
                     catch (const std::invalid_argument&)
                     {
@@ -428,7 +428,7 @@ std::optional<std::string> templateCharReplace(
         try
         {
             size_t pos = 0;
-            int64_t temp = std::stoul(*strPtr, &pos, 0);
+            int64_t temp = std::stoull(*strPtr, &pos, 0);
             if (pos == strPtr->size())
             {
                 keyPair.value() = static_cast<uint64_t>(temp);
